immutable geo index
